ec/google/chromeec: Add shared CFR menu
Modeled after a similar implementation for Lenovo ECs: create a single,
consolidated CFR form for all ChromeEC settings, which is automatically
added to all devices which use ChromeEC and have CFR enabled. Remove
the individual implementations as they are no longer necessary.

Add the SUPPRESS flag for Framework boards as they provide their own
custom CFR options.

TEST=build/boot google mainboards fizz, drobit, morphius.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/95046
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
