+++ src/mainboard/google/dedede/cfr.c
@@ -2,7 +2,6 @@
 
 #include <boot/coreboot_tables.h>
 #include <drivers/option/cfr_frontend.h>
-#include <ec/google/chromeec/cfr.h>
 #include <intelblocks/cfr.h>
 #include <soc/cfr.h>
 
@@ -43,14 +42,6 @@ static struct sm_obj_form system = {
 	},
 };
 
-static struct sm_obj_form ec = {
-	.ui_name = "ChromeEC Embedded Controller",
-	.obj_list = (const struct sm_object *[]) {
-		&ec_kb_backlight,
-		NULL
-	},
-};
-
 static struct sm_obj_form devices = {
 	.ui_name = "Devices",
 	.obj_list = (const struct sm_object *[]) {
@@ -62,7 +53,6 @@ static struct sm_obj_form devices = {
 static struct sm_obj_form *sm_root[] = {
 	&system,
 	&devices,
-	&ec,
 	NULL
 };
