+++ src/mainboard/google/fizz/cfr.c
@@ -3,7 +3,6 @@
 #include <boot/coreboot_tables.h>
 #include <drivers/intel/oc_mailbox/cfr.h>
 #include <drivers/option/cfr_frontend.h>
-#include <ec/google/chromeec/cfr.h>
 #include <intelblocks/cfr.h>
 #include <soc/cfr.h>
 #include "board.h"
@@ -58,14 +57,6 @@ static struct sm_obj_form system = {
 	},
 };
 
-static struct sm_obj_form ec = {
-	.ui_name = "ChromeEC Embedded Controller",
-	.obj_list = (const struct sm_object *[]) {
-		&auto_fan_control,
-		NULL
-	},
-};
-
 static struct sm_obj_form power = {
 	.ui_name = "Power",
 	.obj_list = (const struct sm_object *[]) {
@@ -79,7 +70,6 @@ static struct sm_obj_form power = {
 
 static struct sm_obj_form *sm_root[] = {
 	&system,
-	&ec,
 	&power,
 #if CONFIG(INTEL_OC_MAILBOX_ENABLE_UNDERVOLTING)
 	&cpu_voltage,
