+++ src/mainboard/google/skyrim/Makefile.mk
@@ -8,7 +8,6 @@ ramstage-y += mainboard.c
 ramstage-y += ec.c
 ramstage-y += port_descriptors.c
 ramstage-$(CONFIG_CHROMEOS) += chromeos.c
-ramstage-$(CONFIG_DRIVERS_OPTION_CFR) += cfr.c
 
 verstage-$(CONFIG_VBOOT_STARTS_BEFORE_BOOTBLOCK) += verstage.c
