+++ src/mainboard/google/slippy/cfr.c
@@ -2,7 +2,6 @@
 
 #include <boot/coreboot_tables.h>
 #include <drivers/option/cfr_frontend.h>
-#include <ec/google/chromeec/cfr.h>
 #include <southbridge/intel/lynxpoint/cfr.h>
 
 static const struct sm_object touchpad_type = SM_DECLARE_ENUM({
@@ -45,14 +44,6 @@ static struct sm_obj_form system = {
 	},
 };
 
-static struct sm_obj_form ec = {
-	.ui_name = "ChromeEC Embedded Controller",
-	.obj_list = (const struct sm_object *[]) {
-		&auto_fan_control,
-		NULL
-	},
-};
-
 static struct sm_obj_form devices = {
 	.ui_name = "Devices",
 	.obj_list = (const struct sm_object *[]) {
@@ -66,7 +57,6 @@ static struct sm_obj_form devices = {
 static struct sm_obj_form *sm_root[] = {
 	&system,
 	&devices,
-	&ec,
 	NULL
 };
